[Bug] Change Guild Permissions Update to Replace INTO
Since the guild permissions are not auto populated with defaults (Should they be?), the update query is failing on a clean entry due to the where condition inherit with the update clause.

Swapped over to a replace into format.

Schema Changed to:

CREATE TABLE `guild_permissions` (
	`perm_id` INT(11) NOT NULL DEFAULT '0',
	`guild_id` INT(11) NOT NULL DEFAULT '0',
	`permission` INT(11) NOT NULL DEFAULT '0',
	PRIMARY KEY (`perm_id`, `guild_id`) USING BTREE
)
COLLATE='latin1_swedish_ci'
ENGINE=InnoDB
;

The auto increment `id` field was removed as it was not useful in the current guild permission context. A unique key was created for `perm_id` and `guild_id` as these should be unique combinations.